% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EA_vol.R
\name{calc_EA_vol}
\alias{calc_EA_vol}
\title{FUNCTION_TITLE}
\usage{
calc_EA_vol(smp_mass, cut_vol, massinmachine)
}
\arguments{
\item{smp_mass}{PARAM_DESCRIPTION}

\item{cut_vol}{PARAM_DESCRIPTION}

\item{massinmachine}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
