% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_massinmachine.R
\name{calc_massinmachine}
\alias{calc_massinmachine}
\title{FUNCTION_TITLE}
\usage{
calc_massinmachine(smp_mass, cut_vol, EA_vol, inj_vol)
}
\arguments{
\item{smp_mass}{PARAM_DESCRIPTION}

\item{cut_vol}{PARAM_DESCRIPTION}

\item{EA_vol}{PARAM_DESCRIPTION}

\item{inj_vol}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
