% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_target_mass.R
\name{calc_target_mass}
\alias{calc_target_mass}
\title{FUNCTION_TITLE}
\usage{
calc_target_mass(goal_peak, test_mass, actual_peak)
}
\arguments{
\item{goal_peak}{PARAM_DESCRIPTION}

\item{test_mass}{PARAM_DESCRIPTION}

\item{actual_peak}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
